Exceptions para find e delete

+++ src/main/java/com/desafio/cadastro/services/PessoaService.java
@@ -4,34 +4,46 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import com.desafio.cadastro.entities.Pessoa;
 import com.desafio.cadastro.repositories.PessoaRepository;
+import com.desafio.cadastro.services.exceptions.DatabaseException;
+import com.desafio.cadastro.services.exceptions.ResourceNotFoundException;
 
 @Service
 public class PessoaService {
-	
+
 	@Autowired
 	private PessoaRepository repository;
-	
+
 	public List<Pessoa> findAll() {
 		return repository.findAll();
 	}
-	
+
 	public Pessoa findById(Long id) {
 		Optional<Pessoa> obj = repository.findById(id);
-		return obj.get();
+		return obj.orElseThrow(() -> new ResourceNotFoundException(id));
 	}
-	
+
 	public Pessoa insert(Pessoa obj) {
 		return repository.save(obj);
 	}
-	
+
 	public void delete(Long id) {
-		repository.deleteById(id);
+		try {
+			if(repository.existsById(id) == true) {
+				repository.deleteById(id);				
+			}
+			else {
+				throw new ResourceNotFoundException(id);
+			}
+		}catch(DataIntegrityViolationException e) {
+			throw new DatabaseException(e.getMessage());
+		}
 	}
-	
+
 	public Pessoa update(Long id, Pessoa obj) {
 		Pessoa entity = repository.getReferenceById(id);
 		updateData(entity, obj);
@@ -42,7 +54,6 @@ public class PessoaService {
 		entity.setNome(obj.getNome());
 		entity.setCpf(obj.getCpf());
 		entity.setDataNascimento(obj.getDataNascimento());
-		
+
 	}
 }
-
